pyproject.toml
[project]
name = "ann-stability"
version = "0.1.0"
description = "Stability Study of Sparse, Multi-vector and Filtered ANN Search"
requires-python = ">=3.12"
dependencies = [
    "beir>=2.2.0",
    "datasets>=4.4.1",
    "faiss-cpu>=1.13.1",
    "hnswlib>=0.8.0",
    "matplotlib>=3.10.8",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.2.0",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.3",
]
